[refactor] #92 - auth, login 분리
- auth 는 인증관련 진행이 필요한 작업들이나 회원 탈퇴 등등의 작업 로그인 작업은 아님!
- login이 로그인작업을 진행하는 것임!
