+++ src/main/java/gitbal/backend/api/auth/controller/LoginController.java
@@ -1,0 +1,39 @@
+package gitbal.backend.api.auth.controller;
+
+
+import gitbal.backend.api.auth.service.LoginService;
+import io.jsonwebtoken.io.IOException;
+import io.swagger.v3.oas.annotations.Operation;
+import io.swagger.v3.oas.annotations.responses.ApiResponse;
+import io.swagger.v3.oas.annotations.responses.ApiResponses;
+import io.swagger.v3.oas.annotations.tags.Tag;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
+import lombok.RequiredArgsConstructor;
+import org.springframework.security.core.Authentication;
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestParam;
+
+@Controller
+@RequiredArgsConstructor
+@RequestMapping("/api/v1/login")
+@Tag(name = "로그인 프로세스 관련 API(구현중)", description = "로그인 프로세스에 필요한 api입니다.")
+public class LoginController {
+
+
+    private final LoginService loginService;
+    @GetMapping("/success")
+    @Operation(summary = "로그인 (구현중)", description = "로그인을 위한 api입니다.")
+    @ApiResponses(value = {
+        @ApiResponse(responseCode = "200", description = "로그인에 성공했습니다."),
+    })
+    public String successLogin(@RequestParam String username) throws IOException{
+        return loginService.madeRedirectUrl(username);
+    }
+
+
+
+
+}
